BE-93 Refactor: remove entirely OAuth related imports and naming
