+++ src/main/java/io/openur/domain/user/controller/UserController.java
@@ -9,8 +9,8 @@ import io.openur.domain.user.dto.PatchUserSurveyRequestDto;
 import io.openur.domain.user.exception.UserNotFoundException;
 import io.openur.domain.user.model.Provider;
 import io.openur.domain.user.service.UserService;
-import io.openur.domain.user.service.oauth.LoginService;
-import io.openur.domain.user.service.oauth.LoginServiceFactory;
+import io.openur.domain.user.service.login.LoginService;
+import io.openur.domain.user.service.login.LoginServiceFactory;
 import io.openur.global.common.PagedResponse;
 import io.openur.global.common.Response;
 import io.openur.global.common.UtilController;
