+++ src/main/java/io/openur/domain/user/service/login/LoginService.java
@@ -1,4 +1,4 @@
-package io.openur.domain.user.service.oauth;
+package io.openur.domain.user.service.login;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
